+++ src/TrashLib/Services/Sonarr/ISonarrVersionEnforcement.cs
@@ -1,0 +1,8 @@
+using TrashLib.Services.Sonarr.Config;
+
+namespace Recyclarr.Command;
+
+public interface ISonarrVersionEnforcement
+{
+    Task DoVersionEnforcement(SonarrConfiguration config);
+}